requirements.txt
pydantic>=2.0
pyyaml>=6.0
python-hcl2>=4.0
anthropic>=0.30.0
python-dotenv>=1.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
boto3>=1.34.0
